+++ src/Ampersand/Misc/Options.hs
@@ -76,6 +76,7 @@ data Options = Options { environment :: EnvironmentOptions
                        , test :: Bool
                        , genMetaTables :: Bool -- When set, generate the meta-tables of AST into the prototype
                        , genMetaFile :: Bool  -- When set, output the meta-population as a file
+                       , addMetaRelations :: Bool -- When set, the user can use all relations defined in Formal Ampersand, without the need to specify them explicitly
                        , genRapPopulationOnly :: Bool -- This switch is to tell Ampersand that the model is being used in RAP3 as student's model
                        , sqlHost ::  String  -- do database queries to the specified host
                        , sqlLogin :: String  -- pass login name to the database server
@@ -242,6 +243,7 @@ getOptions' envOpts =
                       , test             = False
                       , genMetaTables    = False
                       , genMetaFile      = False
+                      , addMetaRelations = False
                       , genRapPopulationOnly = False
                       , sqlHost          = "localhost"
                       , sqlLogin         = "ampersand"
@@ -542,6 +544,10 @@ options = [ (Option ['v']   ["version"]
           , (Option []        ["meta-file"]
                (NoArg (\opts -> opts{genMetaFile = True}))
                "Generate the meta-population in AST format and output it to an .adl file"
+            , Hidden)
+          , (Option []        ["add-meta-relations"]
+               (NoArg (\opts -> opts{addMetaRelations = True}))
+               "add the relations of Formal Ampersand into your script"
             , Hidden)
           , (Option []        ["gen-as-rap-model"]
                (NoArg (\opts -> opts{genRapPopulationOnly = True}))