+++ ReleaseNotes.md
@@ -6,6 +6,7 @@
   * Added 'customization' folder to prototype generation process. This folder can be used to e.g. overwrite generated views.
   * Includes for frontend app must be placed in 'app' folder now. The include folder thereby directly matches the destination directory structure.
   * Ampersand version info is printed in verbose mode
+  * New switch, to add formal ampersand relations into your script: --add-meta-relations
 
 ## v3.7.3 (25 november 2016)
   * Alternative definition of univalence and injectivity to get better violations on runtime.